requirements.txt
SQLAlchemy==1.4.22
influxdb==5.3.1
requests==2.25.1
pydantic==1.8.2
Flask==2.0.1
fastapi==0.68.0
python-dotenv==0.19.0
python_geohash==0.8.5
